requirements.txt
Pillow
toml
pyinstaller
python-pptx